Modify Modal url

+++ src/component/userFind/userFind.js
@@ -1,10 +1,13 @@
-import React, { Component } from 'react';
+import React, { Component,Fragment } from 'react';
 import './userFind.scss';
 import {Link} from 'react-router-dom';
 import axios from 'axios';
 import ProjectItem from './proejectItem';
 import {connect} from 'react-redux';
 import Actions from '../../actions/actionType';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faArrowLeft } from '@fortawesome/free-solid-svg-icons'
+import Modal from '../../CommonComponent/Modal/Modal';
 class userFind extends Component {
 
     state = {
@@ -27,27 +30,32 @@ class userFind extends Component {
 
     handleRemove = (index,projectId) => {
         return ()=>{
-            axios.delete(`/api/projectDelete/${projectId}`).then(()=>{
-                this.setState({...this.state,
-                    ProjectList: [...this.state.ProjectList.slice(0,index),...this.state.ProjectList.slice(index+1)]
-                });
-                alert('성공적으로 삭제했습니다.');
-            }).catch((err)=> {alert('삭제에 실패했습니다.');});
+            if(window.confirm('정말로 삭제하시겠습니까?')){
+                axios.delete(`/api/projectDelete/${projectId}`).then(()=>{
+                    this.setState({...this.state,
+                        ProjectList: [...this.state.ProjectList.slice(0,index),...this.state.ProjectList.slice(index+1)]
+                    });
+                    alert('성공적으로 삭제했습니다.');
+                }).catch((err)=> {alert('삭제에 실패했습니다.');});
+            }
         }
     }
 
 
     render() {
-        let a = new Date("2019-01-09T09:11:43.698Z").toUTCString();
         const UserHasProjectList = this.state.ProjectList.map((item,index)=>{
-            const itemDate = new Date(item.createDate).toLocaleString('ko-KR', { timeZone: 'UTC' });
+            const itemDate = new Date(item.createDate).toLocaleString('ko-KR');
             
         return <ProjectItem key={item._id} starNum={item.stars.length} isOwner={this.props.LogInuser === this.props.match.params.userId} projectID={item._id} projectTitle={item.content.Title} handleRemove={this.handleRemove(index,item._id)} 
-        creator={this.props.match.params.userId} createdDate={itemDate}></ProjectItem>})
+        creator={this.props.match.params.userId} createdDate={itemDate} modalTrigger={this.props.modalTrigger}></ProjectItem>})
         return (
+            <Fragment>
             <div className="userFindBox">
                 <div className="userFind__container">
                     <div className="userFind__Header">
+                        <span className='backBtn' onClick={()=>{this.props.history.goBack();}}>
+                            <FontAwesomeIcon icon={faArrowLeft} size="2x"></FontAwesomeIcon>
+                        </span>
                         <Link to='/'>CODE FUN</Link>
                     </div>
                     <div className="userFind__Body">
@@ -62,16 +70,20 @@ class userFind extends Component {
                     </div>
                 </div>
             </div>
+            {this.props.isModalOpen && <Modal></Modal>}
+            </Fragment>
         );
     }
 }
 
 const mapStateToProps = (state) => ({
     LogInuser : state.UserInfo.USER ,
+    isModalOpen: state.Project.Modal.isModalOpen,
 });
 
 const mapDispatchToProps = (dispatch) => ({
-    checkLogin: (token) => dispatch(Actions.checkLoginThunk(token))
+    checkLogin: (token) => dispatch(Actions.checkLoginThunk(token)),
+    modalTrigger: (projectId) => dispatch(Actions.changemodal({ bool: true, component: 'Save', url: `${window.location.hostname}/project/${projectId}`}))
 });
